move index_map to file
